Do not log check_do_not_overlap_failed error, if the cluster is already joined
Two servers could discover each other at the same time and try to join.
One server would join, while another will wait for a lock, get a lock and
print the error in the old version.

In the new version we actually check if it is an overlap, or just fully joined
cluster. In case it is an overlap we still print an error and retry.
In case pids on both remote and local sides are the same, we could just
do nothing (i.e. the cluster looks properly connected at this point).