Do not run check on nodedown in cets_disco
When a node goes down, it would be kicked from other nodes in the cluster
by global. This process in not instant. This process happens even
if node is alive and could reconnect.

Reconnecting too fast would interfere with prevent_overlapped_partitions logic
though.

We still allow for this node to reconnect, but only during the regular checks.
The change temporary puts the node into unavalable_nodes list though